% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ytVideos.R
\name{getVideoStatsDF}
\alias{getVideoStatsDF}
\title{Get statistics for a video on YouTube}
\usage{
getVideoStatsDF(video_id)
}
\arguments{
\item{l}{video_id  Character. Video ID from YouTube. Required.}
}
\value{
A dataframe with multiple variables: \code{id, viewCount, likeCount, dislikeCount, favoriteCount,
commentCount, pullDate}
}
\description{
This function takes a YouTube video ID and returns in a dataframe
statistics like views, likes, and dislikes.
}
