% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ytUtility.R
\name{pasteNA}
\alias{pasteNA}
\title{Paste Columns in a Dataframe and Suppress NAs}
\usage{
pasteNA(..., sep = " ", collapse = NULL, na.rm = F)
}
\arguments{
\item{...}{one or more R objects to be converted to character vectors.}

\item{sep}{String  A character string to separate the terms}

\item{collapse}{String  An optional character string to separate the results.
See the base paste function help for details.  Default is NULL}

\item{na.rm}{Must set to TRUE to suppress NAs}
}
\description{
The function suppresses NAs when pasting columns together
in a data frame.
}
