% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ytComments.R
\name{yt.ChannelComments}
\alias{yt.ChannelComments}
\title{Get Comments from All Videos on a Channel}
\usage{
yt.ChannelComments(channel_id = NULL, published_before = NULL,
  published_after = NULL)
}
\arguments{
\item{channel_id}{String.  The YouTube channel ID.  Cannot be the vanity URL name.}

\item{published_before}{Date.  RFC 339 Format.  Example, "1970-01-01T00:00:00Z"}

\item{published_after}{Date.  RFC 339 Format.  Example, "1970-01-01T00:00:00Z"}
}
\description{
This function collects all the comments from all the videos on a channel,
within a time frame if set.
}
