% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ytAnalytics.R
\name{yt.CommentsOverTime}
\alias{yt.CommentsOverTime}
\title{Graphs of Comments on a Video Over Time}
\usage{
yt.CommentsOverTime(videoComments = NULL, note = NULL)
}
\arguments{
\item{videoComments}{Dataframe object created with the yt.VideoComments() function.
Must have the variables "author_display_name" and "dateTime".}

\item{note}{A string to put a note on the chart, such as the video ID or some other reference.}
}
\value{
Plots two charts, a histogram showing comment density over time and a cumulative line for total
comments over time.
}
\description{
This function produces two charts related to the temporal distribution
of comments on a video over time.  The input to the function is the dataframe
from the yt.VideoComments() function.  The first chart is a daily representation
of number of comments each day, starting with the date of the first comment and ending
on the date of the most recent comment.  The second chart is the cumulative number of
comments on the video over time.
}
