% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ytUtility.R
\name{yt.GetChannelID}
\alias{yt.GetChannelID}
\title{Get YouTube Channel ID from Vanity or Legacy Name}
\usage{
yt.GetChannelID(term = NULL)
}
\arguments{
\item{term}{Vanity name.  String.  Required}
}
\description{
Some YouTube channels have vanity names (sometimes referred to as legacy names)
in the channel URL instead of channel IDs.  The channel ID is needed to collect
any data from the channel, the API will not accept the vanity name.  This function
provides the channel ID for a channel, given the vanity name.
}
