% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ytComments.R
\name{yt.GetCommentReply}
\alias{yt.GetCommentReply}
\title{Get Function for Collecting Comment Replies from a YouTube Video}
\usage{
yt.GetCommentReply(filter = NULL, part = "snippet", text_format = "html",
  max_results = 100, page_token = NULL, ...)
}
\arguments{
\item{filter}{string; Required.
named vector of length 1
potential names of the entry in the vector:
\code{comment_ID}: comment ID.
\code{parent_ID}: parent comment ID.}

\item{part}{Comment resource requested. Required. Comma separated list of one or more of the
following: \code{id, snippet}. e.g., \code{"id, snippet"}, \code{"id"}, etc. Default: \code{snippet}.}

\item{text_format}{Data Type: Character. Default is \code{"html"}. Only takes \code{"html"} or \code{"plainText"}. Optional.}

\item{max_results}{Maximum number of items that should be returned. Integer. Optional. Can be between 20 and 100. Default is 100.}

\item{page_token}{Specific page in the result set that should be returned. Optional.}

\item{\dots}{Additional arguments passed to \code{\link{tuber_GET}}.}
}
\value{
Nested named list.
}
\description{
Basic function, adapted from tuber package, get_comments()
Used by other comment functions in the ytcol package
}
