% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ytComments.R
\name{yt.GetComments}
\alias{yt.GetComments}
\title{Underlying Function to Get Comments on a Video on YouTube
Basic function, adapted from tuber package, get_comment_threads()
Used by other comment functions in the ytcol package}
\usage{
yt.GetComments(filter = NULL, part = "snippet", text_format = "html",
  simplify = TRUE, max_results = 100, page_token = NULL, ...)
}
\arguments{
\item{filter}{string; Required.
named vector of length 1
potential names of the entry in the vector:
\code{video_id}: video ID.
\code{channel_id}: channel ID.
\code{thread_id}: comma-separated list of comment thread IDs
\code{threads_related_to_channel}: channel ID.}

\item{part}{Comment resource requested. Required. Comma separated list of one or more of the
following: \code{id, snippet}. e.g., \code{"id, snippet"}, \code{"id"}, etc. Default: \code{snippet}.}

\item{text_format}{Data Type: Character. Default is \code{"html"}. Only takes \code{"html"} or \code{"plainText"}. Optional.}

\item{simplify}{Data Type: Boolean. Default is \code{TRUE}. If \code{TRUE}, the function returns a data frame. Else a list with all the information returned.}

\item{max_results}{Maximum number of items that should be returned. Integer. Optional. Can be between 20 and 100. Default is 100.}

\item{page_token}{Specific page in the result set that should be returned. Optional.}

\item{\dots}{Additional arguments passed to \code{\link{tuber_GET}}.}
}
\value{
Nested named list.
}
\description{
Underlying Function to Get Comments on a Video on YouTube
Basic function, adapted from tuber package, get_comment_threads()
Used by other comment functions in the ytcol package
}
