% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ytSearch.R
\name{yt.GetSearch}
\alias{yt.GetSearch}
\title{Underlying GET Function for Search}
\usage{
yt.GetSearch(term = term, max_results = max_results,
  published_after = NULL, published_before = NULL, type = "video", ...)
}
\arguments{
\item{term}{Character.  Search term. Required.  No default.}

\item{max_results}{Integer. Maximum number of items returned. Can be between 1 and 50. Default is 50.}

\item{published_after}{Date. Optional. RFC 339 Format.  Example, "1970-01-01T00:00:00Z"}

\item{published_before}{Date. Optional. RFC 339 Format.  Example, "1970-01-01T00:00:00Z"}

\item{type}{Character. Optional. Takes one of three values: \code{'video', 'channel', 'playlist'}. Default is \code{'video'}.}
}
\description{
This function is used by yt.Search in the ytcol package to call the API to get
videos related to a search term.
}
