% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ytAnalytics.R
\name{yt.Network}
\alias{yt.Network}
\title{Network Diagram for Comments on Related Videos}
\usage{
yt.Network(relatedComments = NULL, minVideos = 1)
}
\arguments{
\item{relatedComments}{Dataframe object created with the yt.RelatedVideoComments function.  Must have the variables
"author_display_name" and "video_ID".}

\item{minVideos}{Numeric for the minimum number of videos the author has commented on in the original set of related
videos.  Default is 1. Subsets authors with equal to or greater than the minVideos numeric.}
}
\value{
Plots a network diagram with nodes being comment authors and video IDs.
}
\description{
This function takes the dataframe and generates a network diagram with the nodes being comment authors and videos.
The dataframe must have columns named "comment_ID", "author_display_name", and "video_ID".
The dataframes created with the yt.ChannelComments() and yt.RelatedVideoComments() functions
work well in this function.  The size of the nodes for comment authors indicates the relative number of videos
that author has commented on in the original set of related videos.  The function subsets the
related video comments dataframe by authors that have commented on at least the minimum
number of videos, so as to focus on authors who have commented on multiple videos as more
interesting entities.
}
