% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ytUtility.R
\name{yt.oauth}
\alias{yt.oauth}
\title{Authentication to YouTube}
\usage{
yt.oauth(app_id = NULL, app_secret = NULL, scope = "ssl",
  token = ".httr-oauth", ...)
}
\arguments{
\item{app_id}{client ID.  Required.}

\item{app_secret}{client secret. Required.}

\item{scope}{Character.  Default = "ssl".  Do not change.}

\item{token}{Path to file containing token.}

\item{\dots}{Additional arguments}
}
\description{
The function looks for \code{.httr-oauth} in the working directory. If it doesn't find it, it expects an application ID and a secret.
}
