% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ytVideos.R
\name{yt.related}
\alias{yt.related}
\title{Get Related Videos}
\usage{
yt.related(video_id)
}
\arguments{
\item{video_id}{String.  Video ID from YouTube.}
}
\value{
Dataframe with the following variables: videoID, dateTime, channelID, title, description,
channelTitle, and liveBroadcastContent
}
\description{
Given a video ID, get up to 50 videos that are related.
}
