% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ytVideos.R
\name{yt.RelatedVideoComments}
\alias{yt.RelatedVideoComments}
\title{Get the Comments from Related Videos}
\usage{
yt.RelatedVideoComments(video_id)
}
\arguments{
\item{video_id}{Character.  Video ID from YouTube. Required.}
}
\value{
Dataframe with the following variables describing comments: \code{comment_ID, video_ID, author_display_name,
author_channel_ID, text_display, text_original, dateTime, updated dateTime, reply_count, parent_comment_ID,
parent_author_display_name, parent_author_channel_ID, pullDate, related_to}
}
\description{
Given a video ID, get the comments from that video and videos that are related.
}
