% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ytSearch.R
\name{yt.Search}
\alias{yt.Search}
\title{Search YouTube for a key term with a date range}
\usage{
yt.Search(term = NULL, published_before = NULL, published_after = NULL,
  max_results = 50)
}
\arguments{
\item{term}{Character. Search term. Required.}

\item{published_before}{Date.  Optional. RFC 339 Format.  Example, "1970-01-01T00:00:00Z"}

\item{published_after}{Date.  Optional. RFC 339 Format.  Example, "1970-01-01T00:00:00Z"}

\item{max_results}{Integer.  Maximum number of results returned.  Can be between 1 and 50. Default is 50.}
}
\value{
A dataframe with 6 variables: \code{video_ID, dateTime, channel_ID, title, description, channel_title}
}
\description{
This function makes use of the yt_search function in the tuber package.
and pulls all videos for the set term within the set date range.
}
