% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ytVideos.R
\name{yt.SingleVideoInfo}
\alias{yt.SingleVideoInfo}
\title{Get Statistics and Details for a Single Video on YouTube}
\usage{
yt.SingleVideoInfo(video_id = NULL)
}
\arguments{
\item{video_id}{String.  Video ID from YouTube.}
}
\value{
Dataframe with the following variables: videoID, viewCount, like Count, dislikeCount,
favoriteCount, commentCount, pullDate, dateTime, channelID, title, description, tags,
liveBroadcastContent, localized.title, localized.description
}
\description{
This function combines the stats and details functions
and returns a dataframe with both sets of information for
a single video.  Requires the video ID.
}
