% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ytComments.R
\name{yt.VideoComments}
\alias{yt.VideoComments}
\title{Get Comments on a Video on YouTube}
\usage{
yt.VideoComments(video_id = NULL)
}
\arguments{
\item{video_id}{String.  Video ID from YouTube.}
}
\value{
Dataframe with variables including author information, comment text, date-time posted,
and date-time updated.
}
\description{
This function collects all the comments on a video on YouTube.
It calls the yt.GetComments function in the ytcol package.
}
