% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ytAnalytics.R
\name{yt.Wordcloud}
\alias{yt.Wordcloud}
\title{Wordcloud of Comments from YouTube}
\usage{
yt.Wordcloud(comments = NULL, max_words = 10, remove = NULL,
  stopwords = "english")
}
\arguments{
\item{comments}{Character.  Name of dataframe in the environment that includes the comments in
a column named "text_original".  Required.}

\item{max_words}{Integer.  Number of words to display in the wordcloud.  Default is 10.}

\item{remove}{Comma separated list of words to remove from the words displayed in the wordcloud.  Optional.}

\item{stopwords}{Character.  Stopwords are common words in a language such as “the” and “this”.  Set the langauge
to eliminate common words in that language  Takes one of the following values:
\code{'english','danish','dutch','finnish','french','german','hungarian', 'italian', 'norwegian', 'portuguese',
 'russian', 'spanish', 'swedish'}.  Default is 'english'.}
}
\value{
Plots a wordcloud.
}
\description{
Create a wordcloud from a dataframe that includes the text of comments pulled
from YouTube.  Input dataframe must have a column named "text_original" that contains
the comments.
}
