% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ytUtility.R
\name{yt_check}
\alias{yt_check}
\title{Request Response Verification.
From the tuber package, but not exported.}
\usage{
yt_check(req)
}
\arguments{
\item{req}{Request}
}
\value{
in case of failure, a message
}
\description{
Request Response Verification.
From the tuber package, but not exported.
}
