% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ytUtility.R
\name{yt_check_token}
\alias{yt_check_token}
\title{Check for Authenitication Token in Options.}
\usage{
yt_check_token()
}
\description{
Supports the yt_GET function in the ytcol package.  From the tuber package, original function not exported.
}
