% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ytUtility.R
\name{dataframeFromJSON}
\alias{dataframeFromJSON}
\title{Create a dataframe from JSON}
\usage{
dataframeFromJSON(l)
}
\arguments{
\item{l}{Dataframe$items}
}
\value{
A dataframe of the items
}
\description{
This function converts the JSON provided by the YouTube API to a dataframe.  Utilized
by multiple other functions in the ytcol package.
}
