% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ytUtility.R
\name{dataframeFromJSON}
\alias{dataframeFromJSON}
\title{Create a dataframe from JSON}
\usage{
dataframeFromJSON(l)
}
\arguments{
\item{l}{Dataframe$items from the list_channel_activities()}
}
\value{
A dataframe of the items
}
\description{
This function converts the JSON provided by the YouTube API from the list_channel_activities
function in the tuber package when the part parameter is set to "contentDetails".  The goal is to
capture the video IDs from the channel.
}
