% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ytVideos.R
\name{getVideoDetailsDF}
\alias{getVideoDetailsDF}
\title{Get details for a video on YouTube
This function takes a YouTube video ID and returns details in a dataframe
like date published, channel ID, title, description, channel title, category ID, and tags.}
\usage{
getVideoDetailsDF(video_id)
}
\arguments{
\item{l}{video_id  String.}
}
\value{
A dataframe.
}
\description{
Get details for a video on YouTube
This function takes a YouTube video ID and returns details in a dataframe
like date published, channel ID, title, description, channel title, category ID, and tags.
}
