% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ytVideos.R
\name{getVideoStatsDF}
\alias{getVideoStatsDF}
\title{Get statistics for a video on YouTube
This function takes a YouTube video ID and returns in a dataframe
statistics like views, likes, and dislikes.}
\usage{
getVideoStatsDF(video_id)
}
\arguments{
\item{l}{video_id  String.}
}
\value{
A dataframe.
}
\description{
Get statistics for a video on YouTube
This function takes a YouTube video ID and returns in a dataframe
statistics like views, likes, and dislikes.
}
