% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ytVideos.R
\name{yt.Related}
\alias{yt.Related}
\title{Get Related Videos}
\usage{
yt.Related(video_id)
}
\arguments{
\item{video_id}{Character.  Video ID from YouTube. Required.}
}
\value{
Dataframe with the following variables: \code{video_ID, dateTime, channel_ID, title, description,
channel_title, and related_to}
}
\description{
Given a video ID, get other videos that are related.
}
