% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ytVideos.R
\name{yt.RelatedVideoComments}
\alias{yt.RelatedVideoComments}
\title{Get the Comments from Related Videos}
\usage{
yt.RelatedVideoComments(video_id)
}
\arguments{
\item{video_id}{String.  Video ID from YouTube.}
}
\value{
Dataframe with the following variables describing comments.
}
\description{
Given a video ID, get the comments from that video and videos that are related.
}
