% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ytComments.R
\name{yt.SimpleVideoComments}
\alias{yt.SimpleVideoComments}
\title{Utility for Getting Video Comments}
\usage{
yt.SimpleVideoComments(video_id = NULL)
}
\arguments{
\item{video_id}{Character.  Video ID from YouTube. Required.}
}
\description{
This function is a utility to be used in other functions to get comments for multiple
videos.  Internal use.  Eliminates the save to CSV portion (now eliminated) of the yt.VideoComments function.
}
