% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ytComments.R
\name{yt.VideoComments}
\alias{yt.VideoComments}
\title{Get Comments on a Video on YouTube}
\usage{
yt.VideoComments(video_id = NULL)
}
\arguments{
\item{video_id}{Character.  Video ID from YouTube. Required.}
}
\value{
Dataframe with variables including: \code{comment_ID, video_ID, author_display_name, author_channel_ID,
text_display, text_original, dateTime, updated_dateTime, reply_count, parent_comment_ID, parent_author_display_name,
parent_author_channel_ID, pullDate}
}
\description{
This function collects all the comments on a video on YouTube.
It calls the yt.GetComments function in the ytcol package.
}
