% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ytUtility.R
\name{yt_check}
\alias{yt_check}
\title{Request Response Verification.}
\usage{
yt_check(req)
}
\arguments{
\item{req}{Request}
}
\value{
In case of failure, a message.
}
\description{
Supports the yt_GET function in the ytcol package.  From the tuber package, but not exported.
}
