% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ytUtility.R
\name{yt_check_token}
\alias{yt_check_token}
\title{Check if authenitication token is in options.
From the tuber package, original function not exported.}
\usage{
yt_check_token()
}
\description{
Check if authenitication token is in options.
From the tuber package, original function not exported.
}
