% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ytUtility.R
\name{yt_GET}
\alias{yt_GET}
\title{Base GET function.}
\usage{
yt_GET(path, query, ...)
}
\arguments{
\item{path}{path to specific API request URL}

\item{query}{query list}

\item{\dots}{Additional arguments passed to \code{\link[httr]{GET}}.}
}
\value{
list
}
\description{
From the tuber package, but not exported.
}
